+++ src/climate_learn/models/__init__.py
@@ -1,1 +1,2 @@
 from .module import LitModule
+from .hub import MODEL_REGISTRY
