+++ src/climate_learn/models/hub/__init__.py
@@ -5,3 +5,4 @@ from .persistence import Persistence
 from .resnet import ResNet
 from .unet import Unet
 from .vit import VisionTransformer
+from .utils import MODEL_REGISTRY
