+++ src/climate_learn/models/hub/utils.py
@@ -1,0 +1,7 @@
+MODEL_REGISTRY = {}
+
+def register(name):
+    def decorator(model_class):
+        MODEL_REGISTRY[name] = model_class
+        return model_class
+    return decorator
