+++ src/climate_learn/models/hub/vit.py
@@ -1,21 +1,14 @@
-# Copyright (c) Meta Platforms, Inc. and affiliates.
-# All rights reserved.
+# Local application
+from .components.pos_embed import get_2d_sincos_pos_embed
+from .utils import register
 
-# This source code is licensed under the license found in the
-# LICENSE file in the root directory of this source tree.
-# --------------------------------------------------------
-# References:
-# timm: https://github.com/rwightman/pytorch-image-models/tree/master/timm
-# DeiT: https://github.com/facebookresearch/deit
-# --------------------------------------------------------
-
-
+# Third party
 import torch
 import torch.nn as nn
 from timm.models.vision_transformer import Block, PatchEmbed, trunc_normal_
-from .components.pos_embed import get_2d_sincos_pos_embed
 
 
+@register("vit")
 class VisionTransformer(nn.Module):
     def __init__(
         self,
@@ -167,51 +160,9 @@ class VisionTransformer(nn.Module):
 
         return x
 
-    def forward_loss(
-        self, y, pred, out_variables, metric, lat, log_postfix
-    ):  # metric is a list
-        """
-        y: [N, 3, H, W]
-        pred: [N, L, p*p*3]
-        """
-        pred = self.unpatchify(pred)
-        return (
-            [
-                m(pred, y, out_variables, lat=lat, log_postfix=log_postfix)
-                for m in metric
-            ],
-            pred,
-        )
-
     def forward(self, x, y, out_variables, metric, lat, log_postfix):
         if len(x.shape) == 5:  # history
             x = x.flatten(1, 2)
         embeddings = self.forward_encoder(x)  # B, L, D
         preds = self.head(embeddings)
-        loss, preds = self.forward_loss(
-            y, preds, out_variables, metric, lat, log_postfix
-        )
-        return loss, preds
-
-    def predict(self, x):
-        if len(x.shape) == 5:  # history
-            x = x.flatten(1, 2)
-        with torch.no_grad():
-            embeddings = self.forward_encoder(x)
-            pred = self.head(embeddings)
-        return self.unpatchify(pred)
-
-    def evaluate(
-        self, x, y, variables, out_variables, transform, metrics, lat, clim, log_postfix
-    ):
-        pred = self.predict(x)
-        return [
-            m(pred, y, transform, out_variables, lat, clim, log_postfix)
-            for m in metrics
-        ], pred
-
-
-# model = VisionTransformer(img_size=[32, 64], embed_dim=128, patch_size=2, depth=8, upsampling=2).cuda()
-# x, y = torch.randn(2, 3, 32, 64).cuda(), torch.randn(2, 3, 64, 128).cuda()
-# pred = model.predict(x)
-# print (pred.shape)
+        return self.unpatchify(preds)
