+++ src/climate_learn/models/lr_scheduler.py
@@ -2,12 +2,10 @@ import math
 import warnings
 from typing import List
 
-# from pytorch_lightning.utilities.cli import LR_SCHEDULER_REGISTRY
 from torch.optim import Optimizer
 from torch.optim.lr_scheduler import _LRScheduler
 
 
-# @LR_SCHEDULER_REGISTRY
 class LinearWarmupCosineAnnealingLR(_LRScheduler):
     """Sets the learning rate of each parameter group to follow a linear warmup schedule between
     warmup_start_lr and base_lr followed by a cosine annealing schedule between base_lr and
